fix(elden-ring): map enemy hp as long to match live index (#25)
ensure_index's put_mapping cannot narrow the pre-existing live hp field
(long, from the old NPC-era dynamic mapping) to integer; the reindex aborted
on the first patch. Declare hp as long to match.

+++ packages/elden-ring/elden_ring/_client.py
@@ -235,8 +235,9 @@ INDEX_MAPPING = {
             "menu_category": {"type": "keyword"},
             "npc_id": {"type": "keyword"},
             # Enemy (NpcParam) combat stats — present on the humanoid subset; bosses
-            # are name-only until the chr-id bridge lands (#68).
-            "hp": {"type": "integer"},
+            # are name-only until the chr-id bridge lands (#68). hp is `long` to match
+            # the pre-existing live mapping (put_mapping can't narrow long->integer).
+            "hp": {"type": "long"},
             "stamina": {"type": "integer"},
             "poise": {"type": "float"},
             "magic_defense": {"type": "float"},
